requirements.txt
anthropic
beautifulsoup4
click
cohere
dateparser
google_cloud_aiplatform
langchain_community
llama_index
loguru
openai
pdfplumber
Pillow
protobuf
psycopg2_binary
python_pptx
Requests
striprtf
tiktoken
Unidecode
llama-index-embeddings-langchain
llama-index-vector-stores-postgres